Rename onboarding flow to second pillar flow
