+++ src/entries/components/secondPillarFlow/Steps.scss
