+++ src/entries/components/secondPillarFlow/confirmMandate/ConfirmMandate.js
