+++ src/entries/components/secondPillarFlow/confirmMandate/ConfirmMandate.scss
