+++ src/entries/components/secondPillarFlow/confirmMandate/ConfirmMandate.spec.js
