+++ src/entries/components/secondPillarFlow/confirmMandate/fundTransferTable/FundTransferTable.js
