+++ src/entries/components/secondPillarFlow/confirmMandate/fundTransferTable/FundTransferTable.spec.js
