+++ src/entries/components/secondPillarFlow/confirmMandate/fundTransferTable/index.js
