+++ src/entries/components/secondPillarFlow/confirmMandate/index.js
