+++ src/entries/components/secondPillarFlow/confirmMandate/mandateNotFilledAlert/MandateNotFilledAlert.js
