+++ src/entries/components/secondPillarFlow/confirmMandate/mandateNotFilledAlert/MandateNotFilledAlert.spec.js
