+++ src/entries/components/secondPillarFlow/confirmMandate/mandateNotFilledAlert/index.js
