+++ src/entries/components/secondPillarFlow/selectSources/exactFundSelector/fundExchangeRow/FundExchangeRow.scss
