+++ src/entries/components/secondPillarFlow/selectSources/targetFundSelector/TargetFundSelector.js
