+++ src/entries/components/secondPillarFlow/selectSources/targetFundSelector/TargetFundSelector.scss
