+++ src/entries/components/secondPillarFlow/stepTitle/StepTitle.scss
