+++ src/entries/components/secondPillarFlow/stepTitle/StepTitle.spec.js
