+++ src/entries/components/secondPillarFlow/success/Success.scss
