+++ src/entries/components/secondPillarFlow/transferFutureCapital/targetFundTooltipBody/TargetFundTooltipBody.js
