+++ src/entries/components/secondPillarFlow/transferFutureCapital/targetFundTooltipBody/TargetFundTooltipBody.spec.js
