+++ src/entries/components/secondPillarFlow/transferFutureCapital/targetFundTooltipBody/advanced.png
