+++ src/entries/components/secondPillarFlow/transferFutureCapital/targetFundTooltipBody/conservative.png
