+++ src/entries/components/secondPillarFlow/transferFutureCapital/targetFundTooltipBody/index.js
